refactor(cli): extract research-status command group to cli/ module (#1532 Phase B)
First command-group extraction from the 9.8k-LOC cli.ts god-file. Moves
the 11 read-only research-status commands (objective-state, causal-
trajectory, trust-zone, abstraction-node, cue-anchor, harmonic-search,
and the utility-telemetry / utility-learning family) into
cli/research-status-commands.ts behind a registerResearchStatusCommands
registrar. registerCli now calls the registrar instead of inlining 211
lines of handler code.

Behaviour-preserving by construction: the handler bodies are moved
verbatim; only their nesting (4-space dedent into a function) changes.
The cli-command-surface contract tests (Phase A) stay green unchanged,
guarding against silent drift. The circular import (group module imports
runXxx helpers from ../cli.js; cli.ts imports the registrar) is safe in
ESM — the helpers are referenced only inside action callbacks invoked at
runtime, long after both modules finish evaluating.

Also exports CliApi/CliProgram/CliCommand interfaces so group modules can
type their registrar params, and tightens the dispatch test to exclude
init/migrate (filesystem side effects) from the no-cwd-override loop.

cli.ts: 9872 -> 9663 LOC (ratchet baseline updated downward).
Chokepoints used: runCli in-process entry; per-group registrar pattern.
