+++ scripts/ratchet-baseline.json
@@ -5,7 +5,7 @@
   "metrics": {
     "watchlistLoc": {
       "packages/remnic-core/src/orchestrator.ts": 19870,
-      "packages/remnic-core/src/cli.ts": 9872,
+      "packages/remnic-core/src/cli.ts": 9663,
       "packages/remnic-core/src/access-service.ts": 8566,
       "packages/remnic-core/src/storage.ts": 7686,
       "packages/remnic-core/src/config.ts": 4558
